Fixed RIFT-14957 - RwCType tests fail on Ubuntu
With GCC 5.4 on Ubuntu 16, null check code was optimized out after
call to memcXXX functions.  This resulted in "free" handler called
with null pointer in certain situation and failed regression test
which check for alloc/free balance.

Added check for null pointer before the call to memXXX so that this
extraneous free call is not made.protobuf-c (1.1.0.rw.3-0ubuntu1)
unstable; urgency=medium